+++ packages/dash-platform-queries/src/transition/mod.rs
@@ -1,2 +1,3 @@
 //! Transport-free state transition helpers.
+pub mod put_document;
 pub mod validation;
